+++ python/tests/test_lowlevel.py
@@ -2511,7 +2511,9 @@ class TestTree(LowLevelTestCase):
             st = _tskit.Tree(ts)
             assert st.get_num_nodes() == ts.get_num_nodes()
             # An uninitialised tree should always be zero.
-            assert st.get_left_root() == 0
+            samples = ts.get_samples()
+            assert st.get_left_child(st.get_virtual_root()) == samples[0]
+            assert st.get_right_child(st.get_virtual_root()) == samples[-1]
             assert st.get_left() == 0
             assert st.get_right() == 0
             for j in range(ts.get_num_samples()):
@@ -2656,7 +2658,7 @@ class TestTree(LowLevelTestCase):
         for ts in self.get_example_tree_sequences():
             num_nodes = ts.get_num_nodes()
             st = _tskit.Tree(ts)
-            for v in [num_nodes, 10 ** 6, _tskit.NULL]:
+            for v in [num_nodes + 1, 10 ** 6, _tskit.NULL]:
                 with pytest.raises(ValueError):
                     st.get_mrca(v, v)
                 with pytest.raises(ValueError):
@@ -2694,9 +2696,8 @@ class TestTree(LowLevelTestCase):
             with pytest.raises(ValueError):
                 st.get_newick(root=0, precision=100)
             for precision in range(17):
-                tree = st.get_newick(
-                    root=st.get_left_root(), precision=precision
-                ).decode()
+                root = st.get_left_child(st.get_virtual_root())
+                tree = st.get_newick(root=root, precision=precision).decode()
                 times = get_times(tree)
                 assert len(times) > ts.get_num_samples()
                 for t in times:
@@ -2712,7 +2713,8 @@ class TestTree(LowLevelTestCase):
 
         def check_tree(tree):
             assert tree.get_index() == -1
-            assert tree.get_left_root() == samples[0]
+            assert tree.get_left_child(tree.get_virtual_root()) == samples[0]
+            assert tree.get_num_edges() == 0
             assert tree.get_mrca(0, 1) == _tskit.NULL
             for u in range(ts.get_num_nodes()):
                 assert tree.get_parent(u) == _tskit.NULL
@@ -2739,7 +2741,7 @@ class TestTree(LowLevelTestCase):
             with pytest.raises(TypeError):
                 st.get_newick(ts, buffer_size=bad_type)
         while st.next():
-            u = st.get_left_root()
+            u = st.get_left_child(st.get_virtual_root())
             newick = st.get_newick(u)
             assert newick.endswith(b";")
             with pytest.raises(ValueError):
@@ -2830,7 +2832,7 @@ class TestTree(LowLevelTestCase):
                         assert t.get_left_sample(j) == _tskit.NULL
                         assert t.get_right_sample(j) == _tskit.NULL
                 # The roots should have all samples.
-                u = t.get_left_root()
+                u = t.get_left_child(t.get_virtual_root())
                 samples = []
                 while u != _tskit.NULL:
                     sample = t.get_left_sample(u)
@@ -3032,7 +3034,7 @@ class TestTree(LowLevelTestCase):
         t1.first()
         a = getattr(t1, array + "_array")
         assert a.dtype == np.int32
-        assert a.shape == (ts1.get_num_nodes(),)
+        assert a.shape == (ts1.get_num_nodes() + 1,)
         assert a.base == t1
         assert not a.flags.writeable
         assert a.flags.aligned
@@ -3072,6 +3074,25 @@ class TestTree(LowLevelTestCase):
         # Just do something to touch memory
         a2[:] = 0
         assert a3 is not a2
+
+    @pytest.mark.parametrize("ordering", ["preorder", "postorder"])
+    def test_traversal_arrays(self, ordering):
+        ts = self.get_example_tree_sequence(10)
+        tree = _tskit.Tree(ts)
+        tree.first()
+        method = getattr(tree, "get_" + ordering)
+        for bad_type in [None, {}]:
+            with pytest.raises(TypeError):
+                method(bad_type)
+        for bad_node in [-2, 10 ** 6]:
+            with pytest.raises(_tskit.LibraryError, match="out of bounds"):
+                method(bad_node)
+        a = method(tree.get_virtual_root())
+        assert a.dtype == np.int32
+        assert not a.flags.writeable
+        assert a.flags.aligned
+        assert a.flags.c_contiguous
+        assert a.flags.owndata
 
 
 class TestTableMetadataSchema(MetadataTestMixin):